requirements.txt
loguru
lxml
lxml
pandas
pytest
python-dotenv
Requests
sentry_sdk
